requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
cloudinary==1.41.0
dj-database-url==2.3.0
dj-rest-auth==7.0.0
dj3-cloudinary-storage==0.0.6
Django==5.1.3
django-allauth==65.2.0
django-cloudinary-storage==0.3.0
django-filter==24.3
django-quill-editor==0.1.42
django-rest-auth==0.9.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
idna==3.10
packaging==24.2
pillow==11.0.0
psycopg2==2.9.10
PyJWT==2.10.0
requests==2.32.3
six==1.16.0
sqlparse==0.5.2
typing_extensions==4.12.2
urllib3==2.2.3
whitenoise==6.8.2
